% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_subset.R
\name{qexpression}
\alias{qexpression}
\title{qsubset}
\usage{
qexpression(
  data,
  column = NULL,
  comparitor = "==",
  value = NULL,
  operator = "&"
)
}
\arguments{
\item{data}{the data to subset}

\item{the}{value to compare against}
}
\description{
qsubset
}
\examples{
data(mtcars)
expr <- qexpression(mtcars, "mpg", comparitors$GREATER_THAN, 20)
qsubset(mtcars, expr)
}
