% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_subset.R
\name{qsubset}
\alias{qsubset}
\title{qsubset}
\usage{
qsubset(data, qexpressions)
}
\arguments{
\item{data}{the data to subset}

\item{qexpression}{single qepression or list of qexpressions to subset with}
}
\description{
qsubset
}
\examples{
data(mtcars)
expr <- qexpression(mtcars, "mpg", comparitors$GREATER_THAN, 20)
qsubset(mtcars, expr)
}
