% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_subset.R
\name{subset_ui}
\alias{subset_ui}
\title{Subset UI}
\usage{
subset_ui(input_id, data, steps)
}
\arguments{
\item{input_id}{shiny input id}

\item{steps}{dictionary of steps}

\item{columns}{available columns}
}
\description{
Subset UI
}
\examples{
data(mtcars)
component_example(subset_ui("test", mtcars, list(qexpression(mtcars), qexpression(mtcars))))
}
